+++ src/main/java/nextstep/subway/member/application/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package nextstep.subway.member.application;
 
-import nextstep.subway.auth.ui.UserDetails;
-import nextstep.subway.auth.ui.UserDetailsService;
+import nextstep.subway.auth.infrastructure.UserDetails;
+import nextstep.subway.auth.infrastructure.UserDetailsService;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
