+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,7 +1,7 @@
 package nextstep.subway.member.ui;
 
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
-import nextstep.subway.auth.ui.UserDetails;
+import nextstep.subway.auth.infrastructure.UserDetails;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
